Create Counting_raisins
The image is loaded and converted to a greyscale image and converted to a binary image by thresholding. The highest and the lowest greyscales are found
and then it is averaged, and this value is then used for thresholding. The image obtained is black raisins on a white background and hence, it is inverted
to get white pixels on black background. The image is then filtered using the ‘RAFilter’ function. A 3X3 filter was used as in this case as compared to a 5X5 filter as more border information would be lost. 

The output of this filter is a greyscale and blurry image and hence, it is again thresholded by the middle value, i.e., 125 to remove the blurriness. It is Mid
thresholded as we are only removing only low-level grey values generated by the RAFilter. 

A marker image is generated in which all the pixels excluding 7 pixels from top, bottom, left and right are converted to black (0). This image is then used to remove the boundary pixels by the method ‘Reconstruction by Dilation’. This connects all the boundary raisins which is then subtracted to get the boundary removed raisins image. 8-connected was used instead of 4-connected to increase robustness of the system, as I wanted to consider connectivity in all directions.

Again, a 3X3 RAFilter is used to remove noise for reasons explained above. The ‘Labeller’ function displays individual blobs which have unique greyscale values
for each blob. The centroids of these blobs are identified and given a box around it. 

The number of white pixels is then calculated from the centroid image as the only white pixels left are the centroids, which gives us a perfect count of 17.